fix rollup/parseAst
