+++ packages/client/package.json
@@ -2,9 +2,9 @@
   "name": "client",
   "version": "0.0.0",
   "scripts": {
-    "dev": "tsc --project tsconfig.server.json && cross-env NODE_ENV=development node server/index.js",
+    "dev": "tsc --project tsconfig.server.json && cross-env NODE_ENV=development node server/index.mjs",
     "build": "tsc --project tsconfig.server.json && vite build && vite build --ssr src/entry-server.tsx --outDir dist/server",
-    "preview": "node server/index.js",
+    "preview": "node server/index.mjs",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "format": "prettier --write .",
@@ -42,7 +42,7 @@
     "lefthook": "^1.3.9",
     "prettier": "^3.4.2",
     "ts-jest": "^28.0.8",
-    "typescript": "^4.9.5"
+    "typescript": "^5.8.2"
   },
   "license": "MIT"
 }
