+++ packages/client/server/index.d.mts
@@ -1,0 +1,1 @@
+export {};
