--- packages/client/server/index.d.ts
@@ -1,1 +1,0 @@
-export {}
