+++ packages/client/server/index.mts
@@ -8,7 +8,8 @@ import fs from 'fs/promises'
 import { createServer as createViteServer, ViteDevServer } from 'vite'
 
 const port = process.env.CLIENT_PORT || 3000
-const clientPath = path.join(__dirname, '..')
+const __dirname = path.resolve()
+const clientPath =  __dirname
 const isDev = process.env.NODE_ENV === 'development'
 
 async function createServer() {
