+++ packages/client/src/main.tsx
@@ -1,4 +1,3 @@
-import { BrowserRouter } from 'react-router-dom'
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
