+++ packages/client/tsconfig.server.json
@@ -1,11 +1,13 @@
 {
 	"compilerOptions": {
 		"composite": true,
-		"target": "es2019",
-		"module": "commonjs",
-		"allowSyntheticDefaultImports": true,
-		"esModuleInterop": true,
-		"lib": ["es2019", "DOM"],
+    "target": "ESNext",
+    "module": "ESNext",
+    "moduleResolution": "node",
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
+    "lib": ["es2019", "ESNext", "DOM"],
+    "skipLibCheck": true
 	},
-	"include": ["server/index.ts"]
+	"include": ["server/index.mts"]
 }
